Significant rewrite
I changed function names to clarify and distinguish.  Also group all function that modify the data frame to do so when the data are read in from the file instead of in different functions.  Also, incorporated the DAM number directly into each DAM object so it can be referenced easily.